update: api response dependencies after relocation
